refactor(collections): extract template rendering logic into reusable component
Extract duplicated collection template rendering code from edit and add forms
into a new CollectionTemplateRenderer component. This reduces code duplication
by ~160 lines and centralizes template field rendering logic.

Changes:
- Add CollectionTemplateRenderer component to ~/components/common/index.tsx
- Update edit-entity-collection-information-form.tsx to use new component
- Update add-entity-to-collections-form.tsx to use new component
- Remove duplicate template rendering logic and imports
